Autoloads file for MELPA packaging/compatibility
By adding `sublime-themes-autoloads.el` this directory can now be
packaged automatically by MELPA and can be packaged for Marmalade. While
it serves the same purpose of the ;;;###autoload statements,
`package.el` doesn't currently support parsing files that don't match
the name of the package. Thus, every theme would need to be package
individually. However, I think to reduce the amount of noise (there are
a lot of packages available) it's best to just package all these themes
as a single installable package.