Introduce sublime-themes.el instead of autoloads file
@milkypostman's contribution of the autoloads file is only part of the
story: MELPA needs to grab the package description from somewhere, and
that place should usually be a .el file corresponding to the package
name.

This commit, then, provides a `sublime-themes.el` file with a
description and author info, and it contains an autoloaded block which
will be used to *generate* sublime-theme-autoloads.el at package
installation time.