Add a script to add `themes' folder to theme path.
Theme files are looked up in the directories listed in
`custom-theme-load-path'. I added a script to the root directory called
`emacs-color-themes.el', which adds the full path to themes directory
when loaded to that list.  This means that if the project directory is
in the load path, when I ``(require 'emacs-color-themes)'', I can, e.g.
``(load-theme 'mccarthy t)''.

I did this in to allow to put the whole repo under a directory in the
`load-path' (e.g as a git submodule), so that one can easily use all the
themes, without need to mess with load paths etc.

amend: Update Readme file.